\name{yeast-data}
\docType{data}
\alias{yeast}
\title{A yeast data set for Transcriptional Regulation Inference from Genetics of Gene ExpRession}
\description{
A yeast data set for integrative genomic analysis.
}
\section{Details}{
The data set contains information on 112 F1 segregants from a yeast genetic cross of BY and RM strains.
The list consists of: 
\code{marker:} A 3244 x 112 genotype matrix with marker genotypes in rows and arrays in columns.
\code{exp:} A 6216 x 112 gene expression matrix with genes in rows and arrays in columns.
\code{marker.pos:} A matrix of marker position information.
\code{exp.pos: } A matrix of gene position information.

}
\references{
Brem R.B., Storey J.D., Whittle J., and Kruglyak L. (2005) Genetic interactions between polymorphisms that affect gene expression in yeast. \emph{Nature}, \bold{436(7051):} 701--703.

Storey J.D., Akey J.M., and Kruglyak L. (2005) Multiple locus linkage analysis of genomewide expression in yeast. \emph{PLoS Biology}, \bold{3(8)}: 1380--1390.
}
\keyword{datasets}
